requirements.txt
fastapi
frontend
python-multipart
tools
uvicorn
dotenv
starlette
itsdangerous
pyserial